adjusts "daylight" related functions in "Illum" class to take more daylight types (e.g., D6025)
* changes the input type of "cctToxy(...)" from int to double
* processes user input inside "calDayLightSPD(...)" function including "cct" verification
* updates affected calling functions of these two methods accordingly

+++ lib/rta.h
@@ -129,7 +129,7 @@ namespace rta {
 
             int readSPD( string path, string type );
             void calDayLightSPD( const int cct );
-            vector < double > cctToxy( const int cct ) const;
+            vector < double > cctToxy( const double cct ) const;
 
             const vector < double > getIllumData() const;
             const string getIllumType() const;
